AICP OTA: Don't FC on home builds.
The version checking looks for one of these names
  "EXPERIMENTAL", "RELEASE", "NIGHTLY", "STABLE"

However, a home build might have a name line
  aicp_m7_kitkat-5.0-UNOFFICIAL-20140624

Previously, since "UNOFFICIAL" doesn't match any of the
expected names we tried to parse that as an integer, which
caused an FC. The change was to set the version string
to "" if the name does not match. This prevents
the FC and allows the AICP OTA screen to open. Also,
the index into the list of valid names is left at 0,
which will choose "EXPERIMENTAL" for display purposes.

Change-Id: Icc1f21d62e966bf938e9dc69dba4eedc5f4d6f66