add makeshift legacy update system support
This update client was designed to run on top of the update_engine A/B
update system and file-based encryption. It can't offer the same user
experience and robustness elsewhere. However, due to some recent changes
it's possible to hack in support for the legacy recovery-based update
system. It will handle edge cases like a normal reboot after an update
is downloaded strangely but the basics can work.